Minor changes to style
