+++ src/Dashboard/Dashboard.fxml
@@ -1,14 +1,19 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.scene.image.*?>
-<?import javafx.scene.text.*?>
-<?import java.lang.*?>
-<?import java.util.*?>
-<?import javafx.scene.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
+<?import javafx.scene.canvas.Canvas?>
+<?import javafx.scene.control.DatePicker?>
+<?import javafx.scene.control.SplitPane?>
+<?import javafx.scene.control.Tab?>
+<?import javafx.scene.control.TabPane?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.control.ToggleButton?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Text?>
 
-<SplitPane dividerPositions="0.22742474916387959" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
+<SplitPane dividerPositions="0.22742474916387959" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Dashboard.DashboardController">
   <items>
     <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="160.0" prefWidth="100.0">
          <children>
@@ -16,6 +21,8 @@
                <children>
                   <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Simulation:" />
                   <ToggleButton mnemonicParsing="false" prefHeight="27.0" prefWidth="141.0" text="ON/OFF" />
+                  <ToggleButton mnemonicParsing="false" text="ON/OFF" />
+                  <ToggleButton onAction="#renderLayout" text="Load JSON" />
                   <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Username (or login)" />
                   <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Permisisons:" />
                   <TextArea prefHeight="55.0" prefWidth="132.0" text="-permission 1&#10;-permission 2" />
@@ -35,13 +42,13 @@
          </children></AnchorPane>
     <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="160.0" prefWidth="100.0">
          <children>
-            <SplitPane dividerPositions="0.7626262626262627" layoutX="-1.0" layoutY="2.0" orientation="VERTICAL" prefHeight="398.0" prefWidth="458.0">
+            <SplitPane dividerPositions="0.7626262626262627" layoutX="-1.0" layoutY="2.0" orientation="VERTICAL" prefHeight="598.0" prefWidth="620.0">
               <items>
-                <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="100.0" prefWidth="160.0">
+                <AnchorPane minHeight="-Infinity" minWidth="-Infinity" prefHeight="100.0" prefWidth="618.0">
                      <children>
                         <SplitPane dividerPositions="0.6118421052631579" prefHeight="298.0" prefWidth="458.0">
                           <items>
-                            <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="160.0" prefWidth="100.0">
+                            <AnchorPane minHeight="-Infinity" minWidth="0.0" prefHeight="440.0" prefWidth="275.0">
                                  <children>
                                     <TabPane prefHeight="296.0" prefWidth="275.0" tabClosingPolicy="UNAVAILABLE">
                                       <tabs>
@@ -49,7 +56,7 @@
                                              <content>
                                                 <VBox prefHeight="200.0" prefWidth="100.0">
                                                    <children>
-                                                      <VBox prefHeight="263.0" prefWidth="275.0">
+                                                      <VBox minHeight="-Infinity" prefHeight="406.0" prefWidth="275.0">
                                                          <children>
                                                             <Text strokeType="OUTSIDE" strokeWidth="0.0" />
                                                             <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Modules installed" />
@@ -134,21 +141,22 @@
                                     </TabPane>
                                  </children>
                               </AnchorPane>
-                            <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="160.0" prefWidth="100.0">
+                            <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="440.0" prefWidth="327.0">
                                  <children>
-                                    <ImageView fitHeight="167.0" fitWidth="173.0" layoutY="39.0" pickOnBounds="true" preserveRatio="true">
+                                 <Canvas fx:id="render" height="451.0" width="329.0" />
+                                    <!-- <ImageView fitHeight="167.0" fitWidth="173.0" layoutY="39.0" pickOnBounds="true" preserveRatio="true">
                                        <image>
-                                          <Image url="@../iOS-OS-X-Style-Pure-CSS-Loading-Spinner.jpg" />
-                                       </image>
-                                    </ImageView>
-                                    <Text layoutX="49.0" layoutY="193.0" strokeType="OUTSIDE" strokeWidth="0.0" text="House Layout" />
-                                    <Button layoutX="26.0" layoutY="213.0" mnemonicParsing="false" prefHeight="27.0" prefWidth="122.0" text="Load House File" />
-                                 </children></AnchorPane>
+                                                <Image url="@../iOS-OS-X-Style-Pure-CSS-Loading-Spinner.jpg" />
+                                             </image>
+                                          </ImageView> -->
+                                    <Text layoutX="49.0" layoutY="431.0" strokeType="OUTSIDE" strokeWidth="0.0" text="House Layout" />
+                                 </children>
+                              </AnchorPane>
                           </items>
                         </SplitPane>
                      </children>
                   </AnchorPane>
-                <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="100.0" prefWidth="160.0">
+                <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="288.0" prefWidth="600.0">
                      <children>
                         <Text layoutX="168.0" layoutY="19.0" strokeType="OUTSIDE" strokeWidth="0.0" text="output console" />
                         <TextField layoutX="14.0" layoutY="23.0" prefHeight="64.0" prefWidth="444.0" />
